% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{CalculateRogue}
\alias{CalculateRogue}
\title{ROGUE calculation}
\usage{
CalculateRogue(.x, platform = NULL, cutoff = 0.05, k = NULL,
  features = NULL)
}
\arguments{
\item{.x}{A tibble object returned from the SE_fun or entropy_fit function.}

\item{platform}{The platform ("UMI" or "full-length") used for generating the tested dataset.}

\item{cutoff}{The threshold (adjusted P value) for identifying significant ds. The default threshold is 0.05.}

\item{k}{The scaling factor for calculating ROGUE. The default value of K is set to 45 and 500 for droplet-based ("UMI") and "full-length" based datasets, respectively. When specifying a custom k value, the "platform" argument is redundant.}

\item{features}{Use these features to calculate ROGUE.}
}
\value{
A value of ROGUE.
}
\description{
Using ROGUE to access the purity of single cell population.
}
\details{
By taking advantage of the wide applicability of S-E model to scRNA-seq data, we introduce the statistic ROGUE to measure the purity of a cell population as:

\deqn{ROGUE=1-∑sig.ds/(∑sig.ds+K)}

where K is an important parameter that constrains the ROGUE value between 0 and 1. A cell population with no significant ds for all genes will receive a ROGUE value of 1, while a population with maximum summarization of significant ds is supposed to yield a purity score of ~0.
}
\examples{
ent.res <- SE_fun(expr, span = 0.1, r = 1, mt.method = "fdr")
CalculateRogue(ent.res, platform = "UMI")
CalculateRogue(ent.res, k = 30)
}
