% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{DetermineK}
\alias{DetermineK}
\title{Calculate the value of the reference factor K}
\usage{
DetermineK(expr, span = 0.5, r = 1, mt.method = "fdr", if.adj = T)
}
\arguments{
\item{expr}{The expression matrix. Rows should be genes and columns should be cells.}

\item{span}{The parameter α which controls the degree of smoothing.To improve the performance.}

\item{r}{A small fixed value to avoid log(0) of mean gene expression levels. The default value of r is set to 1, but can also be set to other values such as 0.1 and 0.01.}

\item{mt.method}{The multiple testing method used in p.adjust.}

\item{if.adj}{Whether to apply multiple testing method to adjust p.value.}
}
\value{
The K value
}
\description{
Determine the value of the reference factor K by specifying a highly heterogeneous dataset
}
\examples{
k <- DetermineK(expr, span = 0.5, r = 1, mt.method = "fdr")
}
