% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{entropy_fit}
\alias{entropy_fit}
\title{Fit the relationship between expression entropy and mean gene expression}
\usage{
entropy_fit(.x, span = 0.5, mt.method = c("fdr","BH"))
}
\arguments{
\item{.x}{A tibble object returned from the Entropy function.}

\item{span}{The parameter α which controls the degree of smoothing.}

\item{mt.method}{The multiple testing method used in p.adjust.}
}
\value{
A tibble object with six columns.
}
\description{
Fit the relationship between expression entropy and mean gene expression using loess regression.
}
\examples{
ent.res <- Entropy(expr)
ent.res <- entropy_fit(ent.res, span = 0.3, mt.method = "fdr")
}
