% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{GraphGeodesic}
\alias{GraphGeodesic}
\alias{Cophenetic}
\title{Graph Geodesic between leaves of unweighted tree}
\usage{
GraphGeodesic(x, nTip = length(x$tip.label), log = FALSE, asMatrix = TRUE)

Cophenetic(x, nTip = length(x$tip.label), log = FALSE, asMatrix = TRUE)
}
\arguments{
\item{x}{Object of class \code{phylo}.}

\item{nTip}{Integer specifying number of leaves.}

\item{asMatrix}{Logical specifying whether to coerce output to matrix format.}
}
\value{
\code{GraphGeodesic()} returns an unnamed integer matrix describing the
number of edges between each pair of edges.
}
\description{
Graph Geodesic between leaves of unweighted tree
}
\examples{
GraphGeodesic(TreeTools::BalancedTree(5))
}
\author{
Martin R. Smith, modifying algorithm by Emmanuel Paradis
in \code{ape::dist.nodes()}.
}
\keyword{internal}
