% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{matr.filter}
\alias{matr.filter}
\title{Filtering out low-abundance genes and low-quality cells}
\usage{
matr.filter(expr, min.cells = 10, min.genes = 10)
}
\arguments{
\item{expr}{The expression matrix. Rows should be genes and columns should be cells.}

\item{min.cells}{Include genes detected in at least this many cells.}

\item{min.genes}{Include cells where at least this many genes are detected.}
}
\value{
A filtered gene expression matrix.
}
\description{
Filtering out low-abundance genes and low-quality cells
}
\examples{
expr <- matrix(data = rbinom(n = 100, size = 20, prob = 0.5 ), nrow = 10)
expr
filtered.expr <- matr.filter(expr, min.cells = 3, min.genes = 3)
filtered.expr
}
