% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{rogue.boxplot}
\alias{rogue.boxplot}
\title{Visualize ROGUE values on a boxplot}
\usage{
rogue.boxplot(res.rogue)
}
\arguments{
\item{res.rogue}{A dataframe returned from the 'rogue' function.}
}
\value{
A ggplot object
}
\description{
Draws a boxplot of the ROGUE values for each cluster in different samples.
}
\examples{
res.rogue <- rogue(expr, labels, samples)
rogue.boxplot(res.rogue)
}
