% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RogueNaRok.R
\name{RogueNaRok}
\alias{RogueNaRok}
\title{Call RogueNaRok}
\usage{
RogueNaRok(
  treeFile = "",
  bootTrees = "",
  computeSupport = FALSE,
  run_id = "",
  maxDropsetSize = 0,
  excludeFile = "",
  workdir = "",
  labelPenalty = "",
  mreOptimization = FALSE,
  threshold = 50
)
}
\description{
Call RogueNaRok
}
\author{
\href{https://smithlabdurham.github.io/}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
