% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RogueNaRok.R
\name{Rogues}
\alias{Rogues}
\title{Use RogueNaRok to drop tips and generate more informative consensus}
\usage{
Rogues(
  trees,
  bestTree = NULL,
  computeSupport = TRUE,
  dropsetSize = 1,
  neverDrop = character(0),
  labelPenalty = 0,
  mreOptimization = FALSE,
  threshold = 50
)
}
\arguments{
\item{trees}{List of trees to analyse.}

\item{neverDrop}{Tip labels that should not be dropped from the consensus.}
}
\description{
Use RogueNaRok to drop tips and generate more informative consensus
}
