% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RogueNaRok.R
\name{RogueTaxa}
\alias{RogueTaxa}
\title{Use RogueNaRok to drop tips and generate more informative consensus}
\usage{
RogueTaxa(
  trees,
  bestTree = NULL,
  computeSupport = TRUE,
  dropsetSize = 1,
  neverDrop = character(0),
  labelPenalty = 0,
  mreOptimization = FALSE,
  threshold = 50,
  verbose = FALSE
)
}
\arguments{
\item{trees}{List of trees to analyse.}

\item{neverDrop}{Tip labels that should not be dropped from the consensus.}

\item{verbose}{Logical specifying whether to display output from RogueNaRok.
If \code{FALSE}, output will be included as an attribute of the return value.}
}
\description{
Use RogueNaRok to drop tips and generate more informative consensus
}
