% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGUE.R
\name{SE_fun}
\alias{SE_fun}
\title{Identify highly informative genes using S-E model}
\usage{
SE_fun(expr, span = 0.5, r = 1, mt.method = "fdr", if.adj = T)
}
\arguments{
\item{expr}{The expression matrix. Rows should be genes and columns should be cells.}

\item{span}{The parameter α which controls the degree of smoothing.}

\item{r}{A small fixed value to avoid log(0) of mean gene expression levels. The default value of r is set to 1, but can also be set to other values such as 0.1 and 0.01.}

\item{mt.method}{The multiple testing method used in p.adjust.}

\item{if.adj}{Whether to apply multiple testing method to adjust p.value.}
}
\value{
A tibble object with seven columns:

* Gene, the gene name.

* mean.expr, the mean expression levels of genes.

* entropy, the expected expression entropy from a given mean gene expression.

* fit, the mean expression levels of genes.

* ds, the entropy reduction against the null expectation.

* p.value, the significance of ds.

* p.adj, adjusted P value.
}
\description{
Use S-E curve to identify highly informative genes.
}
\examples{
ent.res <- SE_fun(expr, span = 0.1, r = 1, mt.method = "fdr")
}
