% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIC.R
\name{Cophenetic}
\alias{Cophenetic}
\title{Cophenetic distance between leaves of unweighted tree}
\usage{
Cophenetic(x, nTip = length(x$tip.label))
}
\arguments{
\item{x}{Object of class \code{phylo}.}

\item{nTip}{Integer specifying number of leaves.}
}
\value{
\code{Cophenetic()} returns an unnamed integer matrix describing the
number of edges between each pair of edges.
}
\description{
Cophenetic distance between leaves of unweighted tree
}
\examples{
x <- TreeTools::BalancedTree(6)
}
\author{
Martin R. Smith, modifying algorithm by Emmanuel Paradis
in \code{ape::dist.nodes()}.
}
\keyword{internal}
