% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIC.R
\name{TipInstability}
\alias{TipInstability}
\alias{ColByStability}
\title{Tip instability}
\usage{
TipInstability(trees)

ColByStability(trees, luminence = 50)
}
\arguments{
\item{trees}{List of trees to analyse.}

\item{luminence}{Numeric luminance value to pass to \code{\link[=hcl]{hcl()}}.}
}
\description{
\code{TipInstability()} calculates the instability of each leaf in a tree.
Unstable leaves are likely to display roguish behaviour.
}
\details{
I define the \emph{instability} of a pair of leaves as the median absolute
divergence in the cophenetic distance
(the number of edges in the shortest path between the leaves) across all
trees, normalized against the mean cophenetic distance.
The instability of a single leaf is the mean instability of all pairs that
include that leaf; higher values characterise leaves whose position is more
variable between trees.
}
\examples{
library("TreeTools", quietly = TRUE)
trees <- AddTipEverywhere(BalancedTree(8), 'Rogue')[3:6]
plot(consensus(trees), tip.col = ColByStability(trees))
instab <- TipInstability(trees)
plot(ConsensusWithout(trees, names(instab[instab > 0.2])))
}
\seealso{
Other tip instability functions: 
\code{\link{TipVolatility}()}
}
\author{
\href{https://smithlabdurham.github.io/}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tip instability functions}
